Added: get maximum in float numbers

+++ src/main/java/com/generic/GenericMaximum.java
@@ -1,8 +1,8 @@
 package com.generic;
 
 public class GenericMaximum {
-    public Integer number(Integer firstnumber, Integer secondnumber, Integer thirdnumber){
-        Integer max = firstnumber;
+    public Float number(Float firstnumber, Float secondnumber, Float thirdnumber){
+        Float max = firstnumber;
         if(secondnumber.compareTo(max) > 0)
             max = secondnumber;
         if(thirdnumber.compareTo(max) > 0)
